// site/lib/models/VaccinationInformation.ts
export interface VaccinationInformation {
  name: string
  ci: string
  birthday: string
  departament: string
  municipality: string
  establishment: string
  vaccine: string
  vaccinationDate: string
  dose: string
  supplier: string
  lot: string
  consentNumber: string
  nextVaccinationDate: string
}